pyproject.toml
[project]
requires-python = ">=3.10"
name = "my_project"
version = "0.1.0"
description = "My Python project"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[all]>=0.115.6",
    "python-dotenv>=1.0.1",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn>=0.34.0",
]
